requirements.txt
Flask>=3.0
python-dotenv>=1.0
gunicorn>=21.2
Werkzeug>=3.0

# RAG pipeline
chromadb>=0.5.0
pypdf>=4.0
langchain-text-splitters>=0.3.0

# Ollama client
ollama>=0.3.0

# Rate limiting
flask-limiter>=3.5
